support f16 for quant api

+++ neural_speed/application/common.h
@@ -145,9 +145,9 @@ struct quant_params {
   // [-1, 32, 128]
   int32_t group_size = 32;
   // [fp32, bf16, fp8]
-  std::string scale_dtype = "fp32";
+  std::string scale_dtype = "fp16";
   // [fp32, fp16, bf16, int8]
-  std::string compute_dtype = "fp32";
+  std::string compute_dtype = "int8";
   std::string model_name = "unknown";
   bool use_ggml = false;
   // set by model_name automatically
